introduction

+++ src/Molecule_Lib/iwsubstructure.cc
@@ -1567,7 +1567,8 @@ Single_Substructure_Query::_got_embedding(Query_Atoms_Matched & matched_atoms,
 //  if (results.hits_found() > 40000000)
 //    exit(0);
 
-    if (_max_matches_to_find > 0 && results.hits_found() >= _max_matches_to_find) {
+    if (_max_matches_to_find > 0 &&
+        results.hits_found() >= static_cast<unsigned int>(_max_matches_to_find)) {
       results.set_complete();
       return _max_matches_to_find;
     }
